Move Utils class to package com.cdancy.bitbucket.rest.utils
