+++ src/main/java/com/cdancy/bitbucket/rest/domain/pullrequest/PullRequest.java
@@ -23,8 +23,8 @@ import org.jclouds.javax.annotation.Nullable;
 import org.jclouds.json.SerializedNames;
 
 import com.cdancy.bitbucket.rest.domain.common.ErrorsHolder;
-import com.cdancy.bitbucket.rest.domain.common.Utils;
 import com.cdancy.bitbucket.rest.error.Error;
+import com.cdancy.bitbucket.rest.utils.Utils;
 import com.google.auto.value.AutoValue;
 
 @AutoValue
