+++ src/main/java/com/cdancy/bitbucket/rest/domain/repository/RepositoryPage.java
@@ -24,8 +24,8 @@ import org.jclouds.json.SerializedNames;
 
 import com.cdancy.bitbucket.rest.domain.common.ErrorsHolder;
 import com.cdancy.bitbucket.rest.domain.common.Page;
-import com.cdancy.bitbucket.rest.domain.common.Utils;
 import com.cdancy.bitbucket.rest.error.Error;
+import com.cdancy.bitbucket.rest.utils.Utils;
 import com.google.auto.value.AutoValue;
 
 @AutoValue
